Add z-index to make the hamburger menu float
The addition of a link back to the LU|Zone|UL home on the navbar caused
the action that should have been captured by the hamburger button to
instead by grabbed by the link back to home. Raise the hamburger button
with a little z-index to overcome this in the short term.

Signed-off-by: Dan Scott <dan@coffeecode.net>